+++ src/models/AlbumData.d.ts
@@ -1,11 +1,11 @@
 export default interface AlbumData {
   title: string;
+  shortTitle: string;
   shortDesc: string;
   date: string;
   camera: string;
+  phone: string;
   lenses: string[][];
   previewImgOne: string;
-  previewImgTwo: string;
-  previewImgThree: string;
   content: string;
 }